Continued work on PMS
Right now the Plex Media Server runs all the for the ecosystem on one
host. Eventually I'd like to get this into something like Kubernetes or
something, and distribute it between several worker nodes. For now, I'll
continue configuring Nix to move forward on this one-host direction for
when I can rebuild the server itself.